[ObjLang] Comment examples.
Examples have been commented because they cause index-out-of-bounds errors, which are UB in Lean.
This probably stems from a bug in the definition of object-level inductive types.  While the bug
isn't resolved, they shouldn't be decommented, but are left here as MWE to reproduce the bug.
Since this bug isn't affecting the rest of the project, it is unlikely that it will be solved in
the near future.